C++ MFC Application, reads image and applys different filters ability to change the kernel size  and draw lines on image save & undo buttons
1. Design Overview
The required is to develop a windows based program using C++ and
Microsoft foundation
class [MFC] that can be used to open a Bitmap image and apply some basic
processing on it.
2. Main functions
So the program main functions will be:
1. Open file: a browsing dialog should occur to select the BMP file from
directory and
then display the selected file on your user interface.
2. Apply filer:
It is required to give the user the ability to apply 3 different types
of filters on the
opened Bitmap [Average, Median and Gaussian filters].
The user should have the ability to change the filter size
The user should have the ability to change the Gaussian filter
parameters
3. Draw Lines: give the user a tool to draw lines over the Bmp Using
Device Context
(No OpenGl Here)
4. Undo: the user should have the ability to Undo the applied changes.
5. Save Bitmap: a browsing dialog should occur to select saving
directory and the saved
file name.
3. General Constraints
 The programming language will be Visual C++ [MFC]
 Do not use OpenGl to draw lines